Separate out functions to configure realtime
